Move iter_lstrip to separate utils module
