+++ snekbox/utils/__init__.py
@@ -1,0 +1,3 @@
+from . import iter
+
+__all__ = ("iter",)
